fixing conflicts with master

+++ algorithm.h
@@ -20,7 +20,7 @@ void construct_args(T* p, Args&&... args)
 }
 
 #else
-	
+
 template<typename T, class Arg1> RDE_FORCEINLINE
 void construct_args(T* p, Arg1&& arg1) {
 	::new (static_cast<void*>(p)) T(std::forward<Arg1>(arg1));
@@ -305,11 +305,11 @@ template<class TIter, typename T, class TPred> inline
 TIter lower_bound(TIter first, TIter last, const T& val, const TPred& pred)
 {
 	internal::test_ordering(first, last, pred);
-	int dist(0);
+	ptrdiff_t dist(0);
 	distance(first, last, dist);
 	while (dist > 0)
 	{
-		const int halfDist = dist >> 1;
+		const ptrdiff_t halfDist = dist >> 1;
 		TIter mid = first;
 		rde::advance(mid, halfDist);
 		if (internal::debug_pred(pred, *mid, val))
@@ -325,11 +325,11 @@ template<class TIter, typename T, class TPred> inline
 TIter upper_bound(TIter first, TIter last, const T& val, const TPred& pred)
 {
 	internal::test_ordering(first, last, pred);
-	int dist(0);
+	ptrdiff_t dist(0);
 	distance(first, last, dist);
 	while (dist > 0)
 	{
-		const int halfDist = dist >> 1;
+		const ptrdiff_t halfDist = dist >> 1;
 		TIter mid = first;
 		advance(mid, halfDist);
 		if (!internal::debug_pred(pred, val, *mid))
